Fix most issues when building base as a x64 target.
BUG=1282556

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@112 0039d316-1c4b-4281-b951-d872f2087c98

+++ base/logging.h
@@ -480,7 +480,8 @@ class LogMessage {
 
   LogSeverity severity_;
   std::ostringstream stream_;
-  int message_start_;  // offset of the start of the message (past prefix info).
+  size_t message_start_;  // Offset of the start of the message (past prefix
+                          // info).
 
   DISALLOW_EVIL_CONSTRUCTORS(LogMessage);
 };
